potential for variable type issues

+++ include/project/chassis.h
@@ -9,8 +9,8 @@ class Chassis{
     Chassis();
     void user_control();
     void set(int power,int turn);
-    void move_forward(float distance,int velocity);
-    void point_turn(int angle,int velocity);
+    void move_forward(double distance,int velocity);
+    void point_turn(double angle,int velocity);
 
     double turn_radius;
     double wheel_circumference;
